feat(home): rework featured games section with V2 cover-led card (ARC-747)
Implements the cover-led card design from the FEATURED_GAMES_V2_COVER_LED
handoff: 200px radial cover with a per-game accent, abstract SVG symbol,
genre·pace pill, demo badge, mint solid CTA, and consistent card height.
Pager moved into the section header. Modal updated to drop the legacy
emoji/gradient fields.

Splits apps/web/src/shared/i18n/messages/home.ts into a per-locale folder
(en/es/fr/ru/by) to stay under the 500-line file-length cap.
