+++ apps/web/src/app/[locale]/home/components/HomeGameCard.test.tsx
@@ -24,26 +24,31 @@ vi.mock('next/navigation', () => ({
   useParams: () => ({ locale: 'en' }),
 }));
 
-const mockGame = featuredGames[0]; // critical_v1
+const criticalGame = featuredGames.find((g) => g.id === 'critical_v1')!;
+const glimwormGame = featuredGames.find((g) => g.id === 'glimworm_v1')!;
 
 const baseProps = {
-  game: mockGame,
+  game: criticalGame,
   homeCopy: {
-    gamePlayButton: 'Play Now',
+    gamePlayButton: 'Play now',
+    gameTryDemo: 'Try demo',
     gameComingSoon: 'Coming Soon',
+    gameHowToPlay: 'How to play',
+    gameMetaPlayers: 'players',
+    gameMetaMatch: 'match',
+    gameMetaPlayingNow: 'playing now',
+    demoBadge: 'Demo',
   },
   onOpenDetails: vi.fn(),
 };
 
 describe('HomeGameCard — comingSoon prop', () => {
-  it('renders Play Now button enabled when comingSoon=false', () => {
+  it('renders Play now button enabled when comingSoon=false', () => {
     render(<HomeGameCard {...baseProps} comingSoon={false} />);
 
     const playBtn = screen.getByTestId('game-play-button');
     expect(playBtn).not.toHaveAttribute('aria-disabled', 'true');
-    // href should point to a real path, not '#'
     expect(playBtn.getAttribute('href')).not.toBe('#');
-    // No coming-soon badge
     expect(
       screen.queryByTestId('home-game-coming-soon-badge'),
     ).not.toBeInTheDocument();
@@ -53,28 +58,90 @@ describe('HomeGameCard — comingSoon prop', () => {
     render(<HomeGameCard {...baseProps} comingSoon={true} />);
 
     const playBtn = screen.getByTestId('game-play-button');
-    // Button should be neutralized: href="#" or aria-disabled
     const href = playBtn.getAttribute('href');
     const ariaDisabled = playBtn.getAttribute('aria-disabled');
     expect(href === '#' || ariaDisabled === 'true').toBe(true);
 
-    // Coming-soon badge must be present.
-    // In tests the t() mock returns the key itself ('games.create.comingSoon'),
-    // so we verify the badge exists and its text contains either the real phrase
-    // or the i18n key that represents it.
     const badge = screen.getByTestId('home-game-coming-soon-badge');
     expect(badge).toBeInTheDocument();
     expect(badge.textContent).toMatch(/coming soon|games\.create\.comingSoon/i);
   });
 
   it('still disables when comingSoon=false but isPlayable=false', () => {
-    const nonPlayableGame = { ...mockGame, isPlayable: false };
+    const nonPlayableGame = { ...criticalGame, isPlayable: false };
     render(
       <HomeGameCard {...baseProps} game={nonPlayableGame} comingSoon={false} />,
     );
 
     const playBtn = screen.getByTestId('game-play-button');
-    // Original isPlayable=false path — href must be '#'
     expect(playBtn.getAttribute('href')).toBe('#');
+  });
+});
+
+describe('HomeGameCard — V2 cover-led layout (ARC-747)', () => {
+  it('renders the genre · pace pill in the cover', () => {
+    render(<HomeGameCard {...baseProps} comingSoon={false} />);
+
+    expect(
+      screen.getByText(`${criticalGame.genre} · ${criticalGame.pace}`),
+    ).toBeInTheDocument();
+  });
+
+  it('renders the meta row with players and duration', () => {
+    const { container } = render(
+      <HomeGameCard {...baseProps} comingSoon={false} />,
+    );
+
+    const meta = container.querySelector('.featured-card-meta-main');
+    expect(meta).not.toBeNull();
+    expect(meta!.textContent).toContain(criticalGame.players);
+    expect(meta!.textContent).toContain('players');
+    expect(meta!.textContent).toContain(criticalGame.duration);
+    expect(meta!.textContent).toContain('match');
+  });
+
+  it('shows the Demo badge with "Try demo" CTA copy for demo games', () => {
+    render(<HomeGameCard {...baseProps} game={glimwormGame} />);
+
+    expect(
+      screen.getByTestId(`game-demo-badge-${glimwormGame.id}`),
+    ).toBeInTheDocument();
+    expect(screen.getByTestId('game-play-button').textContent).toMatch(
+      /try demo/i,
+    );
+  });
+
+  it('hides the playing-now meta entry when count is null', () => {
+    render(<HomeGameCard {...baseProps} comingSoon={false} />);
+
+    // criticalGame.playingNow is null in the data
+    expect(screen.queryByText(/playing now/i)).not.toBeInTheDocument();
+  });
+
+  it('shows the playing-now meta entry when count is a number', () => {
+    const liveGame = { ...criticalGame, playingNow: 1234 };
+    render(<HomeGameCard {...baseProps} game={liveGame} />);
+
+    expect(screen.getByText(/1\.2k/)).toBeInTheDocument();
+    expect(screen.getByText(/playing now/i)).toBeInTheDocument();
+  });
+
+  it('exposes the help button with an accessible "How to play" label', () => {
+    render(<HomeGameCard {...baseProps} />);
+
+    const helpBtn = screen.getByTestId('game-help-button');
+    expect(helpBtn).toHaveAccessibleName(/how to play/i);
+  });
+
+  it('applies the per-game accent as a CSS variable', () => {
+    const { container } = render(<HomeGameCard {...baseProps} />);
+
+    const card = container.querySelector(
+      '.featured-card-main',
+    ) as HTMLElement | null;
+    expect(card).not.toBeNull();
+    expect(card!.style.getPropertyValue('--game-accent')).toBe(
+      criticalGame.accentColor,
+    );
   });
 });
