+++ apps/web/src/app/[locale]/home/components/featured-games/symbols/index.ts
@@ -1,0 +1,3 @@
+export * from './CriticalSymbol';
+export * from './SeaBattleSymbol';
+export * from './GlimwormSymbol';
