+++ apps/web/src/app/[locale]/home/data/games.ts
@@ -4,9 +4,24 @@ export interface FeaturedGame {
   id: string;
   nameKey: TranslationKey;
   descriptionKey: TranslationKey;
-  emoji: string;
-  gradient: string;
-  tags: string[];
+  /**
+   * Per-game accent hex used to tint the V2 cover gradient, corner pill,
+   * symbol, and hover border. Falls back to the theme accent if omitted.
+   */
+  accentColor?: string;
+  /** Genre label rendered in the cover corner pill (e.g. "Card", "Board"). */
+  genre: string;
+  /** Pace label rendered alongside the genre (e.g. "Strategy", "Real-time"). */
+  pace: string;
+  /** Player range shown in the meta row (e.g. "3–5"). */
+  players: string;
+  /** Match duration shown in the meta row (e.g. "15 min"). */
+  duration: string;
+  /**
+   * Optional live "playing now" count. Hidden when null/undefined — backend
+   * is expected to populate this later; we currently render a static record.
+   */
+  playingNow?: number | null;
   isPlayable: boolean;
   /** Marks the card with a "DEMO" badge — early preview, not feature-complete. */
   isDemo?: boolean;
@@ -32,9 +47,12 @@ export const featuredGames: FeaturedGame[] = [
     id: 'critical_v1',
     nameKey: 'games.critical_v1.name' as TranslationKey,
     descriptionKey: 'games.critical_v1.description' as TranslationKey,
-    emoji: '💣',
-    gradient: 'linear-gradient(135deg, #FF4D4D 0%, #F9CB28 100%)',
-    tags: ['Card Game', '3-5 Players', '15m Match', 'Strategy'],
+    accentColor: '#f97316',
+    genre: 'Card',
+    pace: 'Strategy',
+    players: '3–5',
+    duration: '15 min',
+    playingNow: null,
     isPlayable: true,
     type: 'card',
     rulesPrefix: 'games.critical_v1.rules',
@@ -53,9 +71,12 @@ export const featuredGames: FeaturedGame[] = [
     id: 'sea_battle_v1',
     nameKey: 'games.sea_battle_v1.name' as TranslationKey,
     descriptionKey: 'games.sea_battle_v1.description' as TranslationKey,
-    emoji: '🚢',
-    gradient: 'linear-gradient(135deg, #3498db 0%, #1abc9c 100%)',
-    tags: ['Board Game', '2-6 Players', '10m Match', 'Strategy'],
+    accentColor: '#38bdf8',
+    genre: 'Board',
+    pace: 'Strategy',
+    players: '2–6',
+    duration: '10 min',
+    playingNow: null,
     isPlayable: true,
     landingHref: '/games/sea-battle',
     type: 'board',
@@ -70,9 +91,12 @@ export const featuredGames: FeaturedGame[] = [
     id: 'glimworm_v1',
     nameKey: 'games.glimworm_v1.name' as TranslationKey,
     descriptionKey: 'games.glimworm_v1.description' as TranslationKey,
-    emoji: '🐛',
-    gradient: 'linear-gradient(135deg, #7c5cff 0%, #5ee0ff 100%)',
-    tags: ['Arcade', '2-10 Players', '90s Match', 'Real-time'],
+    accentColor: '#a78bfa',
+    genre: 'Arcade',
+    pace: 'Real-time',
+    players: '2–10',
+    duration: '90 sec',
+    playingNow: null,
     isPlayable: true,
     isDemo: true,
     type: 'board',
