--- apps/web/src/shared/i18n/messages/home.ts
@@ -1,499 +1,0 @@
-import type { DeepPartial } from '../base-types';
-
-export const en = {
-  kicker: 'Free online board games',
-  tagline:
-    '{{appName}} is the ultimate platform to play free board games online with friends — Battleship, strategy, cards, and more.',
-  description:
-    'Enjoy a wide variety of free board games and tabletop experiences online. Create real-time game rooms, invite your friends, and let {{appName}} handle rules, scoring, and turns so you can focus on the fun.',
-  primaryCtaLabel: 'Get started',
-  playWithBotsLabel: 'Play with Bots',
-  defaultRoomName: "{{name}}'s game",
-  heroCardBrand: 'CRITICAL',
-  heroCardPlayCta: 'Play',
-  supportCtaLabel: 'Support the developers',
-  downloadsTitle: 'Install Arcadeum',
-  pwaDescription:
-    'Install {{appName}} as a Web App on your device for the ultimate board game experience. Same performance, more convenience.',
-  downloadsDescription: 'Grab the latest builds directly from the web app.',
-  downloadsIosLabel: 'Download for iOS',
-  downloadsAndroidLabel: 'Download for Android',
-  gamesTitle: 'Featured Games',
-  gamesSubtitle: 'Explore our growing library of tabletop experiences',
-  gamePlayButton: 'Play Now',
-  gameRulesButton: 'Game Rules',
-  gameComingSoon: 'Coming Soon',
-  gameIncludesPacks: 'Includes 5 Card Packs:',
-  gameThemedDecks: 'Themed Decks:',
-  gameAvailableNow: 'Available Now',
-  showMore: 'Show Details',
-  demoBadge: 'Demo',
-  showLess: 'Show Less',
-  rulesTab: 'Rules',
-  infoTab: 'Game Themes',
-  // Features section
-  featuresTitle: 'Why {{appName}}?',
-  featuresSubtitle:
-    'Everything you need to play board games online with friends',
-  featureRoomsTitle: 'Real-time Rooms',
-  featureRoomsDescription:
-    'Create game rooms instantly and start playing with friends in seconds. No downloads required.',
-  featureRulesTitle: 'Automated Rules',
-  featureRulesDescription:
-    'Let the app handle rules, scoring, and turn management so you can focus on the fun.',
-  featureCrossplatformTitle: 'Cross-platform',
-  featureCrossplatformDescription:
-    'Play instantly in your browser on desktop and mobile. Native apps for iOS and Android coming soon.',
-  featureInviteTitle: 'Private Rooms & Chat',
-  featureInviteDescription:
-    'Create secured rooms for your group with integrated chat to banter while you play.',
-  featureSpectatorTitle: 'Spectator Mode',
-  featureSpectatorDescription:
-    'Watch friends play live on web. TV support coming soon.',
-  featureStatsTitle: 'Game Statistics',
-  featureStatsDescription:
-    'Track your win rates, history, and achievements across all games.',
-  featureTournamentsTitle: 'Tournaments',
-  featureTournamentsDescription:
-    'Compete in ranked events and prove your skills against the best players.',
-  howItWorksTitle: 'How It Works',
-  howItWorksSubtitle: 'Get started in three simple steps',
-  stepCreateTitle: 'Create or Join a Room',
-  stepCreateDescription:
-    'Start a new game room or enter an invite code to join an existing session.',
-  stepInviteTitle: 'Invite Your Friends',
-  stepInviteDescription:
-    'Share the room link or code with friends. They can join from any device instantly.',
-  stepPlayTitle: 'Play Together',
-  stepPlayDescription:
-    'Enjoy board games with {{appName}} handling rules, turns, and scoring automatically.',
-  // Coming soon labels
-  comingSoon: 'Coming Soon',
-  // Footer
-  footerFollowUs: 'Follow Us',
-  footerCommunity: 'Join the community',
-  footerRights: '© {{year}} {{appName}}. All rights reserved.',
-  footerDescription:
-    'Your ultimate destination for competitive and casual gaming experiences.',
-  footerPlatformTitle: 'Platform',
-  footerAllGames: 'All Games',
-  footerSeaBattle: 'Sea Battle',
-  footerTournaments: 'Tournaments',
-  footerLeaderboards: 'Leaderboards',
-  footerRewards: 'Rewards',
-  footerResourcesTitle: 'Resources',
-  footerHelpCenter: 'Help Center',
-  footerGamingBlog: 'Gaming Blog',
-  footerDevelopers: 'Developers (API)',
-  footerLegalTitle: 'Legal',
-  footerPrivacyPolicy: 'Privacy Policy',
-  footerTermsOfService: 'Terms of Service',
-  footerCookiePolicy: 'Cookie Policy',
-  footerContactUs: 'Contact Us',
-  footerStableRelease: 'STABLE RELEASE',
-  footerCraftedWithLove: 'Crafted with passion for gamers worldwide.',
-};
-
-export const es = {
-  kicker: 'Juegos de mesa gratis en línea',
-  tagline:
-    '{{appName}} es tu plataforma para jugar gratis a juegos de mesa en línea con amigos — Batalla Naval, estrategia, cartas y más.',
-  description:
-    'Disfruta de una amplia variedad de juegos de mesa gratis en línea. Crea salas en tiempo real, reúne a tus amigos y deja que {{appName}} automatice reglas, puntuaciones y turnos para que te concentres en la diversión.',
-  primaryCtaLabel: 'Comenzar',
-  playWithBotsLabel: 'Jugar con Bots',
-  defaultRoomName: 'Partida de {{name}}',
-  heroCardBrand: 'CRITICAL',
-  heroCardPlayCta: 'Jugar',
-  supportCtaLabel: 'Apoyar a los desarrolladores',
-  downloadsTitle: 'Instalar Arcadeum',
-  pwaDescription:
-    'Instala {{appName}} como una Web App en tu dispositivo para la mejor experiencia de juego. El mismo rendimiento, más comodidad.',
-  downloadsDescription:
-    'Descarga las últimas versiones directamente desde la web.',
-  downloadsIosLabel: 'Descargar para iOS',
-  downloadsAndroidLabel: 'Descargar para Android',
-  // Games section
-  gamesTitle: 'Juegos Destacados',
-  gamesSubtitle: 'Explora nuestra creciente biblioteca de experiencias de mesa',
-  gamePlayButton: 'Jugar Ahora',
-  gameRulesButton: 'Reglas del Juego',
-  gameComingSoon: 'Próximamente',
-  // Features section
-  featuresTitle: '¿Por qué {{appName}}?',
-  featuresSubtitle:
-    'Todo lo que necesitas para jugar juegos de mesa online con amigos',
-  featureRoomsTitle: 'Salas en Tiempo Real',
-  featureRoomsDescription:
-    'Crea salas de juego al instante y empieza a jugar con amigos en segundos.',
-  featureRulesTitle: 'Reglas Automatizadas',
-  featureRulesDescription:
-    'Deja que la app maneje las reglas, puntuación y turnos para que te concentres en la diversión.',
-  featureCrossplatformTitle: 'Multiplataforma',
-  featureCrossplatformDescription:
-    'Juega en web o móvil. Tus salas se sincronizan perfectamente en todos tus dispositivos.',
-  featureInviteTitle: 'Invita a Amigos',
-  featureInviteDescription:
-    'Comparte códigos de invitación o enlaces para reunir a tus amigos desde cualquier lugar.',
-  featureSpectatorTitle: 'Modo Espectador',
-  featureSpectatorDescription:
-    'Mira a tus amigos jugar en vivo en la web. Soporte para TV próximamente.',
-  featureStatsTitle: 'Estadísticas de Juego',
-  featureStatsDescription:
-    'Sigue tus tasas de victoria, historial y logros en todos los juegos.',
-  featureTournamentsTitle: 'Torneos',
-  featureTournamentsDescription:
-    'Compite en eventos clasificados y demuestra tus habilidades contra los mejores.',
-  // Games Section Extras
-  gameIncludesPacks: 'Incluye 5 Paquetes de Cartas:',
-  gameThemedDecks: 'Mazos Temáticos:',
-  gameAvailableNow: 'Disponible Ahora',
-  showMore: 'Ver Detalles',
-  demoBadge: 'Demo',
-  showLess: 'Ver Menos',
-  rulesTab: 'Reglas',
-  infoTab: 'Temas del Juego',
-  // ... how it works ...
-  howItWorksTitle: 'Cómo Funciona',
-  howItWorksSubtitle: 'Comienza en tres simples pasos',
-  stepCreateTitle: 'Crea o Únete a una Sala',
-  stepCreateDescription:
-    'Inicia una nueva sala de juego o introduce un código de invitación para unirte a una sesión existente.',
-  stepInviteTitle: 'Invita a tus Amigos',
-  stepInviteDescription:
-    'Comparte el enlace o código de la sala con amigos. Puedен unirse desde cualquier dispositivo al instante.',
-  stepPlayTitle: 'Juega Juntos',
-  stepPlayDescription:
-    'Disfruta de juegos de mesa mientras {{appName}} maneja reglas, turnos y puntuación automáticamente.',
-  // Coming soon labels
-  comingSoon: 'Próximamente',
-  mobileComingSoonDescription:
-    'Las apps nativas de iOS y Android para {{appName}} están en desarrollo. ¡Mantente atento!',
-  // Footer
-  footerFollowUs: 'Síguenos',
-  footerCommunity: 'Únete a la comunidad',
-  footerRights: '© {{year}} {{appName}}. Todos los derechos reservados.',
-  footerDescription:
-    'Tu destino definitivo para experiencias de juego competitivas y casuales.',
-  footerPlatformTitle: 'Plataforma',
-  footerAllGames: 'Todos los juegos',
-  footerSeaBattle: 'Batalla Naval',
-  footerTournaments: 'Torneos',
-  footerLeaderboards: 'Tablas de clasificación',
-  footerRewards: 'Recompensas',
-  footerResourcesTitle: 'Recursos',
-  footerHelpCenter: 'Centro de ayuda',
-  footerGamingBlog: 'Blog de juegos',
-  footerDevelopers: 'Desarrolladores (API)',
-  footerLegalTitle: 'Legal',
-  footerPrivacyPolicy: 'Política de privacidad',
-  footerTermsOfService: 'Términos de servicio',
-  footerCookiePolicy: 'Política de cookies',
-  footerContactUs: 'Contáctenos',
-  footerStableRelease: 'VERSIÓN ESTABLE',
-  footerCraftedWithLove: 'Creado con pasión para jugadores de todo el mundo.',
-};
-
-export const fr = {
-  kicker: 'Jeux de société gratuits en ligne',
-  tagline:
-    '{{appName}} est votre plateforme pour jouer gratuitement à des jeux de société en ligne entre amis — Bataille navale, stratégie, cartes et plus.',
-  description:
-    "Profitez d'une large sélection de jeux de société gratuits en ligne. Créez des salons en temps réel, rassemblez vos amis et laissez {{appName}} automatiser règles, scores et tours pour que vous puissiez vous concentrer sur le plaisir.",
-  primaryCtaLabel: 'Commencer',
-  playWithBotsLabel: 'Jouer avec des Bots',
-  defaultRoomName: 'Partie de {{name}}',
-  heroCardBrand: 'CRITICAL',
-  heroCardPlayCta: 'Jouer',
-  supportCtaLabel: 'Soutenir les développeurs',
-  downloadsTitle: 'Installer Arcadeum',
-  pwaDescription:
-    'Installez {{appName}} en tant que Web App sur votre appareil pour une expérience de jeu ultime. Mêmes performances, plus de commodité.',
-  downloadsDescription:
-    'Téléchargez les dernières versions directement depuis le web.',
-  downloadsIosLabel: 'Télécharger pour iOS',
-  downloadsAndroidLabel: 'Télécharger pour Android',
-  // Games section
-  gamesTitle: 'Jeux en Vedette',
-  gamesSubtitle:
-    "Explorez notre bibliothèque croissante d'expériences de table",
-  gamePlayButton: 'Jouer Maintenant',
-  gameRulesButton: 'Règles du Jeu',
-  gameComingSoon: 'Bientôt Disponible',
-  // Features section
-  featuresTitle: 'Pourquoi {{appName}} ?',
-  featuresSubtitle:
-    'Tout ce dont vous avez besoin pour jouer aux jeux de société en ligne with vos amis',
-  featureRoomsTitle: 'Salons en Temps Réel',
-  featureRoomsDescription:
-    'Créez des salons de jeu instantanément et commencez à jouer avec vos amis en quelques secondes.',
-  featureRulesTitle: 'Règles Automatisées',
-  featureRulesDescription:
-    "Laissez l'application gérer les règles, les scores et les tours pour vous concentrer sur le plaisir.",
-  featureCrossplatformTitle: 'Multiplateforme',
-  featureCrossplatformDescription:
-    'Jouez instantanément dans votre navigateur sur bureau et mobile. Applications natives pour iOS et Android bientôt disponibles.',
-  featureInviteTitle: 'Salons Privés et Chat',
-  featureInviteDescription:
-    'Créez des salons sécurisés pour votre groupe avec chat intégré pour discuter pendant que vous jouez.',
-  featureSpectatorTitle: 'Mode Spectateur',
-  featureSpectatorDescription:
-    'Regardez vos amis jouer en direct sur le web. Support TV bientôt disponible.',
-  featureStatsTitle: 'Statistiques de Jeu',
-  featureStatsDescription:
-    'Suivez vos taux de victoire, votre historique et vos réalisations dans tous les jeux.',
-  featureTournamentsTitle: 'Tournois',
-  featureTournamentsDescription:
-    'Participez à des événements classés et prouvez vos compétences contre les meilleurs joueurs.',
-  // Games Section Extras
-  gameIncludesPacks: 'Inclus 5 Paquets de Cartes :',
-  gameThemedDecks: 'Decks Thématiques :',
-  gameAvailableNow: 'Disponible Maintenant',
-  showMore: 'Voir les Détails',
-  demoBadge: 'Démo',
-  showLess: 'Voir Moins',
-  rulesTab: 'Reglas',
-  infoTab: 'Thèmes du Jeu',
-  // How it works section
-  howItWorksTitle: 'Comment Ça Marche',
-  howItWorksSubtitle: 'Commencez en trois étapes simples',
-  stepCreateTitle: 'Créez ou Rejoignez un Salon',
-  stepCreateDescription:
-    "Démarrez un nouveau salon de jeu ou entrez un code d'invitation pour rejoindre une session existante.",
-  stepInviteTitle: 'Invitez Vos Amis',
-  stepInviteDescription:
-    "Partagez le lien ou le code du salon avec vos amis. Ils peuvent rejoindre depuis n'importe quel appareil instantanément.",
-  stepPlayTitle: 'Jouez Ensemble',
-  stepPlayDescription:
-    'Profitez des jeux de société pendant que {{appName}} gère les règles, les tours et les scores automatiquement.',
-  // Coming soon labels
-  comingSoon: 'Bientôt disponible',
-  mobileComingSoonDescription:
-    "Les applications natives iOS et Android pour {{appName}} sont en développement. Restez à l'écoute !",
-  // Footer
-  footerFollowUs: 'Suivez-nous',
-  footerCommunity: 'Rejoindre la communauté',
-  footerRights: '© {{year}} {{appName}}. Tous droits réservés.',
-  footerDescription:
-    'Votre destination ultime pour des expériences de jeu compétitives et décontractées.',
-  footerPlatformTitle: 'Plateforme',
-  footerAllGames: 'Tous les jeux',
-  footerSeaBattle: 'Bataille Navale',
-  footerTournaments: 'Tournois',
-  footerLeaderboards: 'Classements',
-  footerRewards: 'Récompenses',
-  footerResourcesTitle: 'Ressources',
-  footerHelpCenter: "Centre d'aide",
-  footerGamingBlog: 'Blog de jeu',
-  footerDevelopers: 'Développeurs (API)',
-  footerLegalTitle: 'Légal',
-  footerPrivacyPolicy: 'Politique de confidentialité',
-  footerTermsOfService: "Conditions d'utilisation",
-  footerCookiePolicy: 'Politique relative aux cookies',
-  footerContactUs: 'Contactez-nous',
-  footerStableRelease: 'VERSION STABLE',
-  footerCraftedWithLove: 'Créé avec passion pour les joueurs du monde entier.',
-};
-
-export const ru = {
-  kicker: 'Бесплатные настольные игры онлайн',
-  tagline:
-    '{{appName}} — платформа для бесплатной игры в настольные игры онлайн с друзьями: Морской бой, стратегии, карты и больше.',
-  description:
-    'Большой выбор бесплатных настольных игр онлайн. Создавайте игровые комнаты в реальном времени, приглашайте друзей, а {{appName}} возьмет на себя правила, подсчет очков и очередность ходов, чтобы вы могли сосредоточиться на веселье.',
-  primaryCtaLabel: 'Начать',
-  playWithBotsLabel: 'Играть с ботами',
-  defaultRoomName: 'Игра {{name}}',
-  heroCardBrand: 'CRITICAL',
-  heroCardPlayCta: 'Играть',
-  supportCtaLabel: 'Поддержать разработчиков',
-  downloadsTitle: 'Установить Arcadeum',
-  pwaDescription:
-    'Установите {{appName}} как веб-приложение на свое устройство для лучшего игрового опыта. Та же производительность, больше удобства.',
-  downloadsDescription: 'Загрузите последние версии прямо из веб-приложения.',
-  downloadsIosLabel: 'Скачать для iOS',
-  downloadsAndroidLabel: 'Скачать для Android',
-  // Games section
-  gamesTitle: 'Популярные игры',
-  gamesSubtitle: 'Исследуйте нашу растущую библиотеку настольных развлечений',
-  gamePlayButton: 'Играть сейчас',
-  gameRulesButton: 'Правила игры',
-  gameComingSoon: 'Скоро',
-  gameIncludesPacks: 'Включает 5 наборов карт:',
-  gameThemedDecks: 'Тематические колоды:',
-  gameAvailableNow: 'Доступно сейчас',
-  showMore: 'Подробнее',
-  demoBadge: 'Демо',
-  showLess: 'Свернуть',
-  rulesTab: 'Правила',
-  infoTab: 'Темы игры',
-
-  // Features section
-  featuresTitle: 'Почему {{appName}}?',
-  featuresSubtitle: 'Все, что нужно для игры в настолки онлайн с друзьями',
-  featureRoomsTitle: 'Комнаты в реальном времени',
-  featureRoomsDescription:
-    'Создавайте игровые комнаты мгновенно и начинайте играть с друзьями за секунды. Загрузка не требуется.',
-  featureRulesTitle: 'Автоматизированные правила',
-  featureRulesDescription:
-    'Пусть приложение берет на себя правила, очки и управление ходами, пока вы наслаждаетесь игрой.',
-  featureCrossplatformTitle: 'Кроссплатформенность',
-  featureCrossplatformDescription:
-    'Играйте мгновенно в браузере на ПК и мобильных устройствах. Нативные приложения для iOS and Android скоро появятся.',
-  featureInviteTitle: 'Приватные комнаты и чат',
-  featureInviteDescription:
-    'Создавайте защищенные комнаты для своей группы со встроенным чатом для общения во время игры.',
-  featureSpectatorTitle: 'Режим зрителя',
-  featureSpectatorDescription:
-    'Смотрите, как друзья играют в прямом эфире на вебе. Поддержка ТВ скоро появится.',
-  featureStatsTitle: 'Статистика игры',
-  featureStatsDescription:
-    'Отслеживайте процент побед, историю и достижения во всех играх.',
-  featureTournamentsTitle: 'Турниры',
-  featureTournamentsDescription:
-    'Участвуйте в рейтинговых событиях и докажите свое мастерство против лучших игроков.',
-  // How it works section
-  howItWorksTitle: 'Как это работает',
-  howItWorksSubtitle: 'Начните за три простых шага',
-  stepCreateTitle: 'Создайте или войдите в комнату',
-  stepCreateDescription:
-    'Начните новую игровую комнату или введите код приглашения, чтобы присоединиться к существующей сессии.',
-  stepInviteTitle: 'Пригласите друзей',
-  stepInviteDescription:
-    'Поделитесь ссылкой на комнату или кодом с друзьями. Они могут присоединиться с любого устройства мгновенно.',
-  stepPlayTitle: 'Играйте вместе',
-  stepPlayDescription:
-    'Наслаждайтесь настольными играми, пока {{appName}} автоматически управляет правилами, ходами и очками.',
-  // Coming soon labels
-  comingSoon: 'Скоро',
-  // Footer
-  footerFollowUs: 'Подпишитесь на нас',
-  footerCommunity: 'Присоединяйтесь к сообществу',
-  footerRights: '© {{year}} {{appName}}. Все права защищены.',
-  footerDescription:
-    'Ваша идеальная платформа для соревновательных и казуальных настольных игр.',
-  footerPlatformTitle: 'Платформа',
-  footerAllGames: 'Все игры',
-  footerSeaBattle: 'Морской бой',
-  footerTournaments: 'Турниры',
-  footerLeaderboards: 'Таблицы лидеров',
-  footerRewards: 'Награды',
-  footerResourcesTitle: 'Ресурсы',
-  footerHelpCenter: 'Центр помощи',
-  footerGamingBlog: 'Блог о играх',
-  footerDevelopers: 'Разработчикам (API)',
-  footerLegalTitle: 'Юридическая информация',
-  footerPrivacyPolicy: 'Политика конфиденциальности',
-  footerTermsOfService: 'Условия использования',
-  footerCookiePolicy: 'Политика использования файлов cookie',
-  footerContactUs: 'Связаться с нами',
-  footerStableRelease: 'СТАБИЛЬНЫЙ ВЫПУСК',
-  footerCraftedWithLove: 'Создано с любовью для игроков со всего мира.',
-};
-
-export const by = {
-  kicker: 'Бясплатныя настольныя гульні анлайн',
-  tagline:
-    '{{appName}} — платформа для бясплатнай гульні ў настольныя гульні анлайн з сябрамі: Марскі бой, стратэгіі, карты і больш.',
-  description:
-    'Шырокі выбар бясплатных настольных гульняў анлайн. Стварайце гульнявыя пакоі ў рэальным часе, запрашайце сяброў, а {{appName}} возьме на сябе правілы, падлік ачкоў і чарговасць хадоў, каб вы маглі засяродзіцца на весялосці.',
-  primaryCtaLabel: 'Пачаць',
-  playWithBotsLabel: 'Гуляць з ботамі',
-  defaultRoomName: 'Гульня {{name}}',
-  heroCardBrand: 'CRITICAL',
-  heroCardPlayCta: 'Гуляць',
-  supportCtaLabel: 'Падрымаць распрацоўшчыкаў',
-  downloadsTitle: 'Усталяваць Arcadeum',
-  pwaDescription:
-    'Усталюйце {{appName}} як вэб-прыкладанне на сваю прыладу для лепшага гульнявога досведу. Тая ж прадукцыйнасць, больш зручнасці.',
-  downloadsDescription: 'Загрузіце апошнія версіі прама з вэб-прыкладання.',
-  downloadsIosLabel: 'Спампаваць для iOS',
-  downloadsAndroidLabel: 'Спампаваць для Android',
-  // Games section
-  gamesTitle: 'Папулярныя гульні',
-  gamesSubtitle: 'Даследуйце нашу бібліятэку настольных забаў, якая расце',
-  gamePlayButton: 'Гуляць зараз',
-  gameRulesButton: 'Правілы гульні',
-  gameComingSoon: 'Хутка',
-  gameIncludesPacks: 'Уключае 5 набораў карт:',
-  gameThemedDecks: 'Тэматычныя калоды:',
-  gameAvailableNow: 'Даступна зараз',
-  showMore: 'Падрабязней',
-  demoBadge: 'Дэма',
-  showLess: 'Згарнуць',
-  rulesTab: 'Правілы',
-  infoTab: 'Тэмы гульні',
-
-  // Features section
-  featuresTitle: 'Чаму {{appName}}?',
-  featuresSubtitle: 'Усё, што трэба для гульні ў настолкі онлайн з сябрамі',
-  featureRoomsTitle: 'Пакоі ў рэальным часе',
-  featureRoomsDescription:
-    'Стварайце гульнявыя пакоі імгненна і пачынайце гуляць з сябрамі за секунды. Загрузка не патрабуецца.',
-  featureRulesTitle: 'Аўтаматызаваныя правілы',
-  featureRulesDescription:
-    'Няхай прыкладанне бярэ на сябе правілы, ачкі і кіраванне хадамі, пакуль вы атрымліваеце асалоду ад гульні.',
-  featureCrossplatformTitle: 'Кроссплатформавасць',
-  featureCrossplatformDescription:
-    'Гуляйце імгненна ў браўзеры на ПК і мабільных прыладах. Натыўныя прыкладанні для iOS and Android хутка з’явыцца.',
-  featureInviteTitle: 'Прыватныя пакоі і чат',
-  featureInviteDescription:
-    'Стварайце абароненыя пакоі для сваёй групы з убудаваным чатам для камунікацыі падчас гульні.',
-  featureSpectatorTitle: 'Рэжым гледача',
-  featureSpectatorDescription:
-    'Глядзіце, как сябры гуляюць у прамым эфіры на вэбе. Падтрымка ТБ хутка з’явіцца.',
-  featureStatsTitle: 'Статыстыка гульні',
-  featureStatsDescription:
-    'Сачыце за працэнтам перамог, гісторыяй і дасягненнямі ва ўсіх гульнях.',
-  featureTournamentsTitle: 'Турніры',
-  featureTournamentsDescription:
-    'Удзельнічайце ў рэйтынгавых падзеях і дакажыце сваё майстэрства супраць лепшых гульцоў.',
-  // How it works section
-  howItWorksTitle: 'Як гэта працуе',
-  howItWorksSubtitle: 'Пачніце за тры простыя крокі',
-  stepCreateTitle: 'Стварыце або ўвайдзіце ў пакой',
-  stepCreateDescription:
-    'Пачніце новы гульнявы пакой або ўвядзіце код запрашэння, каб далучыцца да існуючай сесіі.',
-  stepInviteTitle: 'Запрасіце сяброў',
-  stepInviteDescription:
-    'Падзяліцеся спасылкай на пакой або кодам з сябрамі. Яны могуць далучыцца з любой прылады імгненна.',
-  stepPlayTitle: 'Гуляць разам',
-  stepPlayDescription:
-    'Атрымлівайце асалоду ад настольных гульняў, пакуль {{appName}} аўтаматычна кіруе правіламі, хадамі і ачкамі.',
-  // Coming soon labels
-  comingSoon: 'Хутка',
-  // Footer
-  footerFollowUs: 'Падпішыцеся на нас',
-  footerCommunity: 'Далучайцеся да супольнасці',
-  footerRights: '© {{year}} {{appName}}. Усе правы абаронены.',
-  footerDescription:
-    'Ваша ідэальная платформа для спаборніцкіх і казуальных настольных гульняў.',
-  footerPlatformTitle: 'Платформа',
-  footerAllGames: 'Усе гульні',
-  footerSeaBattle: 'Марскі бой',
-  footerTournaments: 'Турніры',
-  footerLeaderboards: 'Табліцы лідараў',
-  footerRewards: 'Узнагароды',
-  footerResourcesTitle: 'Рэсурсы',
-  footerHelpCenter: 'Цэнтр дапамогі',
-  footerGamingBlog: 'Блог пра гульні',
-  footerDevelopers: 'Распрацоўшчыкам (API)',
-  footerLegalTitle: 'Юрыдычная інфармацыя',
-  footerPrivacyPolicy: 'Палітыка прыватнасці',
-  footerTermsOfService: 'Умовы карыстання',
-  footerCookiePolicy: 'Паліка выкарыстання файлаў cookie',
-  footerContactUs: 'Звязацца з намі',
-  footerStableRelease: 'СТАБІЛЬНЫ ВЫПУСК',
-  footerCraftedWithLove: 'Створана з любоўю для гульцоў з усяго свету.',
-};
-
-export const homeMessages = {
-  en,
-  es,
-  fr,
-  ru,
-  by,
-} as const;
-
-/** Derived type with Partial wrapper for backward compatibility */
-export type HomeMessages = DeepPartial<typeof en>;
