+++ apps/web/src/shared/i18n/messages/home/by.ts
@@ -1,0 +1,104 @@
+export const by = {
+  kicker: 'Бясплатныя настольныя гульні анлайн',
+  tagline:
+    '{{appName}} — платформа для бясплатнай гульні ў настольныя гульні анлайн з сябрамі: Марскі бой, стратэгіі, карты і больш.',
+  description:
+    'Шырокі выбар бясплатных настольных гульняў анлайн. Стварайце гульнявыя пакоі ў рэальным часе, запрашайце сяброў, а {{appName}} возьме на сябе правілы, падлік ачкоў і чарговасць хадоў, каб вы маглі засяродзіцца на весялосці.',
+  primaryCtaLabel: 'Пачаць',
+  playWithBotsLabel: 'Гуляць з ботамі',
+  defaultRoomName: 'Гульня {{name}}',
+  heroCardBrand: 'CRITICAL',
+  heroCardPlayCta: 'Гуляць',
+  supportCtaLabel: 'Падрымаць распрацоўшчыкаў',
+  downloadsTitle: 'Усталяваць Arcadeum',
+  pwaDescription:
+    'Усталюйце {{appName}} як вэб-прыкладанне на сваю прыладу для лепшага гульнявога досведу. Тая ж прадукцыйнасць, больш зручнасці.',
+  downloadsDescription: 'Загрузіце апошнія версіі прама з вэб-прыкладання.',
+  downloadsIosLabel: 'Спампаваць для iOS',
+  downloadsAndroidLabel: 'Спампаваць для Android',
+  // Games section
+  gamesTitle: 'Папулярныя гульні',
+  gamesSubtitle: 'Даследуйце нашу бібліятэку настольных забаў, якая расце',
+  gamePlayButton: 'Гуляць зараз',
+  gameTryDemo: 'Паспрабаваць дэма',
+  gameRulesButton: 'Правілы гульні',
+  gameComingSoon: 'Хутка',
+  gameHowToPlay: 'Як гуляць',
+  gameMetaPlayers: 'гульцоў',
+  gameMetaMatch: 'партыя',
+  gameMetaPlayingNow: 'гуляюць зараз',
+  gamesPagerLabel: 'Карусель папулярных гульняў',
+  gamesPagerPrev: 'Папярэдняя гульня',
+  gamesPagerNext: 'Наступная гульня',
+  gameIncludesPacks: 'Уключае 5 набораў карт:',
+  gameThemedDecks: 'Тэматычныя калоды:',
+  gameAvailableNow: 'Даступна зараз',
+  showMore: 'Падрабязней',
+  demoBadge: 'Дэма',
+  demoBadgeLabel: 'Дэма-зборка',
+  showLess: 'Згарнуць',
+  rulesTab: 'Правілы',
+  infoTab: 'Тэмы гульні',
+
+  // Features section
+  featuresTitle: 'Чаму {{appName}}?',
+  featuresSubtitle: 'Усё, што трэба для гульні ў настолкі онлайн з сябрамі',
+  featureRoomsTitle: 'Пакоі ў рэальным часе',
+  featureRoomsDescription:
+    'Стварайце гульнявыя пакоі імгненна і пачынайце гуляць з сябрамі за секунды. Загрузка не патрабуецца.',
+  featureRulesTitle: 'Аўтаматызаваныя правілы',
+  featureRulesDescription:
+    'Няхай прыкладанне бярэ на сябе правілы, ачкі і кіраванне хадамі, пакуль вы атрымліваеце асалоду ад гульні.',
+  featureCrossplatformTitle: 'Кроссплатформавасць',
+  featureCrossplatformDescription:
+    'Гуляйце імгненна ў браўзеры на ПК і мабільных прыладах. Натыўныя прыкладанні для iOS and Android хутка з’явыцца.',
+  featureInviteTitle: 'Прыватныя пакоі і чат',
+  featureInviteDescription:
+    'Стварайце абароненыя пакоі для сваёй групы з убудаваным чатам для камунікацыі падчас гульні.',
+  featureSpectatorTitle: 'Рэжым гледача',
+  featureSpectatorDescription:
+    'Глядзіце, как сябры гуляюць у прамым эфіры на вэбе. Падтрымка ТБ хутка з’явіцца.',
+  featureStatsTitle: 'Статыстыка гульні',
+  featureStatsDescription:
+    'Сачыце за працэнтам перамог, гісторыяй і дасягненнямі ва ўсіх гульнях.',
+  featureTournamentsTitle: 'Турніры',
+  featureTournamentsDescription:
+    'Удзельнічайце ў рэйтынгавых падзеях і дакажыце сваё майстэрства супраць лепшых гульцоў.',
+  // How it works section
+  howItWorksTitle: 'Як гэта працуе',
+  howItWorksSubtitle: 'Пачніце за тры простыя крокі',
+  stepCreateTitle: 'Стварыце або ўвайдзіце ў пакой',
+  stepCreateDescription:
+    'Пачніце новы гульнявы пакой або ўвядзіце код запрашэння, каб далучыцца да існуючай сесіі.',
+  stepInviteTitle: 'Запрасіце сяброў',
+  stepInviteDescription:
+    'Падзяліцеся спасылкай на пакой або кодам з сябрамі. Яны могуць далучыцца з любой прылады імгненна.',
+  stepPlayTitle: 'Гуляць разам',
+  stepPlayDescription:
+    'Атрымлівайце асалоду ад настольных гульняў, пакуль {{appName}} аўтаматычна кіруе правіламі, хадамі і ачкамі.',
+  // Coming soon labels
+  comingSoon: 'Хутка',
+  // Footer
+  footerFollowUs: 'Падпішыцеся на нас',
+  footerCommunity: 'Далучайцеся да супольнасці',
+  footerRights: '© {{year}} {{appName}}. Усе правы абаронены.',
+  footerDescription:
+    'Ваша ідэальная платформа для спаборніцкіх і казуальных настольных гульняў.',
+  footerPlatformTitle: 'Платформа',
+  footerAllGames: 'Усе гульні',
+  footerSeaBattle: 'Марскі бой',
+  footerTournaments: 'Турніры',
+  footerLeaderboards: 'Табліцы лідараў',
+  footerRewards: 'Узнагароды',
+  footerResourcesTitle: 'Рэсурсы',
+  footerHelpCenter: 'Цэнтр дапамогі',
+  footerGamingBlog: 'Блог пра гульні',
+  footerDevelopers: 'Распрацоўшчыкам (API)',
+  footerLegalTitle: 'Юрыдычная інфармацыя',
+  footerPrivacyPolicy: 'Палітыка прыватнасці',
+  footerTermsOfService: 'Умовы карыстання',
+  footerCookiePolicy: 'Паліка выкарыстання файлаў cookie',
+  footerContactUs: 'Звязацца з намі',
+  footerStableRelease: 'СТАБІЛЬНЫ ВЫПУСК',
+  footerCraftedWithLove: 'Створана з любоўю для гульцоў з усяго свету.',
+};
