+++ apps/web/src/shared/i18n/messages/home/en.ts
@@ -1,0 +1,102 @@
+export const en = {
+  kicker: 'Free online board games',
+  tagline:
+    '{{appName}} is the ultimate platform to play free board games online with friends — Battleship, strategy, cards, and more.',
+  description:
+    'Enjoy a wide variety of free board games and tabletop experiences online. Create real-time game rooms, invite your friends, and let {{appName}} handle rules, scoring, and turns so you can focus on the fun.',
+  primaryCtaLabel: 'Get started',
+  playWithBotsLabel: 'Play with Bots',
+  defaultRoomName: "{{name}}'s game",
+  heroCardBrand: 'CRITICAL',
+  heroCardPlayCta: 'Play',
+  supportCtaLabel: 'Support the developers',
+  downloadsTitle: 'Install Arcadeum',
+  pwaDescription:
+    'Install {{appName}} as a Web App on your device for the ultimate board game experience. Same performance, more convenience.',
+  downloadsDescription: 'Grab the latest builds directly from the web app.',
+  downloadsIosLabel: 'Download for iOS',
+  downloadsAndroidLabel: 'Download for Android',
+  gamesTitle: 'Featured Games',
+  gamesSubtitle: 'Explore our growing library of tabletop experiences',
+  gamePlayButton: 'Play now',
+  gameTryDemo: 'Try demo',
+  gameRulesButton: 'Game Rules',
+  gameComingSoon: 'Coming Soon',
+  gameHowToPlay: 'How to play',
+  gameMetaPlayers: 'players',
+  gameMetaMatch: 'match',
+  gameMetaPlayingNow: 'playing now',
+  gamesPagerLabel: 'Featured games carousel',
+  gamesPagerPrev: 'Previous game',
+  gamesPagerNext: 'Next game',
+  gameIncludesPacks: 'Includes 5 Card Packs:',
+  gameThemedDecks: 'Themed Decks:',
+  gameAvailableNow: 'Available Now',
+  showMore: 'Show Details',
+  demoBadge: 'Demo',
+  demoBadgeLabel: 'Demo build',
+  showLess: 'Show Less',
+  rulesTab: 'Rules',
+  infoTab: 'Game Themes',
+  // Features section
+  featuresTitle: 'Why {{appName}}?',
+  featuresSubtitle:
+    'Everything you need to play board games online with friends',
+  featureRoomsTitle: 'Real-time Rooms',
+  featureRoomsDescription:
+    'Create game rooms instantly and start playing with friends in seconds. No downloads required.',
+  featureRulesTitle: 'Automated Rules',
+  featureRulesDescription:
+    'Let the app handle rules, scoring, and turn management so you can focus on the fun.',
+  featureCrossplatformTitle: 'Cross-platform',
+  featureCrossplatformDescription:
+    'Play instantly in your browser on desktop and mobile. Native apps for iOS and Android coming soon.',
+  featureInviteTitle: 'Private Rooms & Chat',
+  featureInviteDescription:
+    'Create secured rooms for your group with integrated chat to banter while you play.',
+  featureSpectatorTitle: 'Spectator Mode',
+  featureSpectatorDescription:
+    'Watch friends play live on web. TV support coming soon.',
+  featureStatsTitle: 'Game Statistics',
+  featureStatsDescription:
+    'Track your win rates, history, and achievements across all games.',
+  featureTournamentsTitle: 'Tournaments',
+  featureTournamentsDescription:
+    'Compete in ranked events and prove your skills against the best players.',
+  howItWorksTitle: 'How It Works',
+  howItWorksSubtitle: 'Get started in three simple steps',
+  stepCreateTitle: 'Create or Join a Room',
+  stepCreateDescription:
+    'Start a new game room or enter an invite code to join an existing session.',
+  stepInviteTitle: 'Invite Your Friends',
+  stepInviteDescription:
+    'Share the room link or code with friends. They can join from any device instantly.',
+  stepPlayTitle: 'Play Together',
+  stepPlayDescription:
+    'Enjoy board games with {{appName}} handling rules, turns, and scoring automatically.',
+  // Coming soon labels
+  comingSoon: 'Coming Soon',
+  // Footer
+  footerFollowUs: 'Follow Us',
+  footerCommunity: 'Join the community',
+  footerRights: '© {{year}} {{appName}}. All rights reserved.',
+  footerDescription:
+    'Your ultimate destination for competitive and casual gaming experiences.',
+  footerPlatformTitle: 'Platform',
+  footerAllGames: 'All Games',
+  footerSeaBattle: 'Sea Battle',
+  footerTournaments: 'Tournaments',
+  footerLeaderboards: 'Leaderboards',
+  footerRewards: 'Rewards',
+  footerResourcesTitle: 'Resources',
+  footerHelpCenter: 'Help Center',
+  footerGamingBlog: 'Gaming Blog',
+  footerDevelopers: 'Developers (API)',
+  footerLegalTitle: 'Legal',
+  footerPrivacyPolicy: 'Privacy Policy',
+  footerTermsOfService: 'Terms of Service',
+  footerCookiePolicy: 'Cookie Policy',
+  footerContactUs: 'Contact Us',
+  footerStableRelease: 'STABLE RELEASE',
+  footerCraftedWithLove: 'Crafted with passion for gamers worldwide.',
+};
