+++ apps/web/src/shared/i18n/messages/home/es.ts
@@ -1,0 +1,108 @@
+export const es = {
+  kicker: 'Juegos de mesa gratis en línea',
+  tagline:
+    '{{appName}} es tu plataforma para jugar gratis a juegos de mesa en línea con amigos — Batalla Naval, estrategia, cartas y más.',
+  description:
+    'Disfruta de una amplia variedad de juegos de mesa gratis en línea. Crea salas en tiempo real, reúne a tus amigos y deja que {{appName}} automatice reglas, puntuaciones y turnos para que te concentres en la diversión.',
+  primaryCtaLabel: 'Comenzar',
+  playWithBotsLabel: 'Jugar con Bots',
+  defaultRoomName: 'Partida de {{name}}',
+  heroCardBrand: 'CRITICAL',
+  heroCardPlayCta: 'Jugar',
+  supportCtaLabel: 'Apoyar a los desarrolladores',
+  downloadsTitle: 'Instalar Arcadeum',
+  pwaDescription:
+    'Instala {{appName}} como una Web App en tu dispositivo para la mejor experiencia de juego. El mismo rendimiento, más comodidad.',
+  downloadsDescription:
+    'Descarga las últimas versiones directamente desde la web.',
+  downloadsIosLabel: 'Descargar para iOS',
+  downloadsAndroidLabel: 'Descargar para Android',
+  // Games section
+  gamesTitle: 'Juegos Destacados',
+  gamesSubtitle: 'Explora nuestra creciente biblioteca de experiencias de mesa',
+  gamePlayButton: 'Jugar Ahora',
+  gameTryDemo: 'Probar demo',
+  gameRulesButton: 'Reglas del Juego',
+  gameComingSoon: 'Próximamente',
+  gameHowToPlay: 'Cómo jugar',
+  gameMetaPlayers: 'jugadores',
+  gameMetaMatch: 'partida',
+  gameMetaPlayingNow: 'jugando ahora',
+  gamesPagerLabel: 'Carrusel de juegos destacados',
+  gamesPagerPrev: 'Juego anterior',
+  gamesPagerNext: 'Siguiente juego',
+  demoBadgeLabel: 'Versión demo',
+  // Features section
+  featuresTitle: '¿Por qué {{appName}}?',
+  featuresSubtitle:
+    'Todo lo que necesitas para jugar juegos de mesa online con amigos',
+  featureRoomsTitle: 'Salas en Tiempo Real',
+  featureRoomsDescription:
+    'Crea salas de juego al instante y empieza a jugar con amigos en segundos.',
+  featureRulesTitle: 'Reglas Automatizadas',
+  featureRulesDescription:
+    'Deja que la app maneje las reglas, puntuación y turnos para que te concentres en la diversión.',
+  featureCrossplatformTitle: 'Multiplataforma',
+  featureCrossplatformDescription:
+    'Juega en web o móvil. Tus salas se sincronizan perfectamente en todos tus dispositivos.',
+  featureInviteTitle: 'Invita a Amigos',
+  featureInviteDescription:
+    'Comparte códigos de invitación o enlaces para reunir a tus amigos desde cualquier lugar.',
+  featureSpectatorTitle: 'Modo Espectador',
+  featureSpectatorDescription:
+    'Mira a tus amigos jugar en vivo en la web. Soporte para TV próximamente.',
+  featureStatsTitle: 'Estadísticas de Juego',
+  featureStatsDescription:
+    'Sigue tus tasas de victoria, historial y logros en todos los juegos.',
+  featureTournamentsTitle: 'Torneos',
+  featureTournamentsDescription:
+    'Compite en eventos clasificados y demuestra tus habilidades contra los mejores.',
+  // Games Section Extras
+  gameIncludesPacks: 'Incluye 5 Paquetes de Cartas:',
+  gameThemedDecks: 'Mazos Temáticos:',
+  gameAvailableNow: 'Disponible Ahora',
+  showMore: 'Ver Detalles',
+  demoBadge: 'Demo',
+  showLess: 'Ver Menos',
+  rulesTab: 'Reglas',
+  infoTab: 'Temas del Juego',
+  // ... how it works ...
+  howItWorksTitle: 'Cómo Funciona',
+  howItWorksSubtitle: 'Comienza en tres simples pasos',
+  stepCreateTitle: 'Crea o Únete a una Sala',
+  stepCreateDescription:
+    'Inicia una nueva sala de juego o introduce un código de invitación para unirte a una sesión existente.',
+  stepInviteTitle: 'Invita a tus Amigos',
+  stepInviteDescription:
+    'Comparte el enlace o código de la sala con amigos. Puedен unirse desde cualquier dispositivo al instante.',
+  stepPlayTitle: 'Juega Juntos',
+  stepPlayDescription:
+    'Disfruta de juegos de mesa mientras {{appName}} maneja reglas, turnos y puntuación automáticamente.',
+  // Coming soon labels
+  comingSoon: 'Próximamente',
+  mobileComingSoonDescription:
+    'Las apps nativas de iOS y Android para {{appName}} están en desarrollo. ¡Mantente atento!',
+  // Footer
+  footerFollowUs: 'Síguenos',
+  footerCommunity: 'Únete a la comunidad',
+  footerRights: '© {{year}} {{appName}}. Todos los derechos reservados.',
+  footerDescription:
+    'Tu destino definitivo para experiencias de juego competitivas y casuales.',
+  footerPlatformTitle: 'Plataforma',
+  footerAllGames: 'Todos los juegos',
+  footerSeaBattle: 'Batalla Naval',
+  footerTournaments: 'Torneos',
+  footerLeaderboards: 'Tablas de clasificación',
+  footerRewards: 'Recompensas',
+  footerResourcesTitle: 'Recursos',
+  footerHelpCenter: 'Centro de ayuda',
+  footerGamingBlog: 'Blog de juegos',
+  footerDevelopers: 'Desarrolladores (API)',
+  footerLegalTitle: 'Legal',
+  footerPrivacyPolicy: 'Política de privacidad',
+  footerTermsOfService: 'Términos de servicio',
+  footerCookiePolicy: 'Política de cookies',
+  footerContactUs: 'Contáctenos',
+  footerStableRelease: 'VERSIÓN ESTABLE',
+  footerCraftedWithLove: 'Creado con pasión para jugadores de todo el mundo.',
+};
