+++ apps/web/src/shared/i18n/messages/home/fr.ts
@@ -1,0 +1,109 @@
+export const fr = {
+  kicker: 'Jeux de société gratuits en ligne',
+  tagline:
+    '{{appName}} est votre plateforme pour jouer gratuitement à des jeux de société en ligne entre amis — Bataille navale, stratégie, cartes et plus.',
+  description:
+    "Profitez d'une large sélection de jeux de société gratuits en ligne. Créez des salons en temps réel, rassemblez vos amis et laissez {{appName}} automatiser règles, scores et tours pour que vous puissiez vous concentrer sur le plaisir.",
+  primaryCtaLabel: 'Commencer',
+  playWithBotsLabel: 'Jouer avec des Bots',
+  defaultRoomName: 'Partie de {{name}}',
+  heroCardBrand: 'CRITICAL',
+  heroCardPlayCta: 'Jouer',
+  supportCtaLabel: 'Soutenir les développeurs',
+  downloadsTitle: 'Installer Arcadeum',
+  pwaDescription:
+    'Installez {{appName}} en tant que Web App sur votre appareil pour une expérience de jeu ultime. Mêmes performances, plus de commodité.',
+  downloadsDescription:
+    'Téléchargez les dernières versions directement depuis le web.',
+  downloadsIosLabel: 'Télécharger pour iOS',
+  downloadsAndroidLabel: 'Télécharger pour Android',
+  // Games section
+  gamesTitle: 'Jeux en Vedette',
+  gamesSubtitle:
+    "Explorez notre bibliothèque croissante d'expériences de table",
+  gamePlayButton: 'Jouer Maintenant',
+  gameTryDemo: 'Essayer la démo',
+  gameRulesButton: 'Règles du Jeu',
+  gameComingSoon: 'Bientôt Disponible',
+  gameHowToPlay: 'Comment jouer',
+  gameMetaPlayers: 'joueurs',
+  gameMetaMatch: 'partie',
+  gameMetaPlayingNow: 'en train de jouer',
+  gamesPagerLabel: 'Carrousel des jeux en vedette',
+  gamesPagerPrev: 'Jeu précédent',
+  gamesPagerNext: 'Jeu suivant',
+  demoBadgeLabel: 'Version démo',
+  // Features section
+  featuresTitle: 'Pourquoi {{appName}} ?',
+  featuresSubtitle:
+    'Tout ce dont vous avez besoin pour jouer aux jeux de société en ligne with vos amis',
+  featureRoomsTitle: 'Salons en Temps Réel',
+  featureRoomsDescription:
+    'Créez des salons de jeu instantanément et commencez à jouer avec vos amis en quelques secondes.',
+  featureRulesTitle: 'Règles Automatisées',
+  featureRulesDescription:
+    "Laissez l'application gérer les règles, les scores et les tours pour vous concentrer sur le plaisir.",
+  featureCrossplatformTitle: 'Multiplateforme',
+  featureCrossplatformDescription:
+    'Jouez instantanément dans votre navigateur sur bureau et mobile. Applications natives pour iOS et Android bientôt disponibles.',
+  featureInviteTitle: 'Salons Privés et Chat',
+  featureInviteDescription:
+    'Créez des salons sécurisés pour votre groupe avec chat intégré pour discuter pendant que vous jouez.',
+  featureSpectatorTitle: 'Mode Spectateur',
+  featureSpectatorDescription:
+    'Regardez vos amis jouer en direct sur le web. Support TV bientôt disponible.',
+  featureStatsTitle: 'Statistiques de Jeu',
+  featureStatsDescription:
+    'Suivez vos taux de victoire, votre historique et vos réalisations dans tous les jeux.',
+  featureTournamentsTitle: 'Tournois',
+  featureTournamentsDescription:
+    'Participez à des événements classés et prouvez vos compétences contre les meilleurs joueurs.',
+  // Games Section Extras
+  gameIncludesPacks: 'Inclus 5 Paquets de Cartes :',
+  gameThemedDecks: 'Decks Thématiques :',
+  gameAvailableNow: 'Disponible Maintenant',
+  showMore: 'Voir les Détails',
+  demoBadge: 'Démo',
+  showLess: 'Voir Moins',
+  rulesTab: 'Reglas',
+  infoTab: 'Thèmes du Jeu',
+  // How it works section
+  howItWorksTitle: 'Comment Ça Marche',
+  howItWorksSubtitle: 'Commencez en trois étapes simples',
+  stepCreateTitle: 'Créez ou Rejoignez un Salon',
+  stepCreateDescription:
+    "Démarrez un nouveau salon de jeu ou entrez un code d'invitation pour rejoindre une session existante.",
+  stepInviteTitle: 'Invitez Vos Amis',
+  stepInviteDescription:
+    "Partagez le lien ou le code du salon avec vos amis. Ils peuvent rejoindre depuis n'importe quel appareil instantanément.",
+  stepPlayTitle: 'Jouez Ensemble',
+  stepPlayDescription:
+    'Profitez des jeux de société pendant que {{appName}} gère les règles, les tours et les scores automatiquement.',
+  // Coming soon labels
+  comingSoon: 'Bientôt disponible',
+  mobileComingSoonDescription:
+    "Les applications natives iOS et Android pour {{appName}} sont en développement. Restez à l'écoute !",
+  // Footer
+  footerFollowUs: 'Suivez-nous',
+  footerCommunity: 'Rejoindre la communauté',
+  footerRights: '© {{year}} {{appName}}. Tous droits réservés.',
+  footerDescription:
+    'Votre destination ultime pour des expériences de jeu compétitives et décontractées.',
+  footerPlatformTitle: 'Plateforme',
+  footerAllGames: 'Tous les jeux',
+  footerSeaBattle: 'Bataille Navale',
+  footerTournaments: 'Tournois',
+  footerLeaderboards: 'Classements',
+  footerRewards: 'Récompenses',
+  footerResourcesTitle: 'Ressources',
+  footerHelpCenter: "Centre d'aide",
+  footerGamingBlog: 'Blog de jeu',
+  footerDevelopers: 'Développeurs (API)',
+  footerLegalTitle: 'Légal',
+  footerPrivacyPolicy: 'Politique de confidentialité',
+  footerTermsOfService: "Conditions d'utilisation",
+  footerCookiePolicy: 'Politique relative aux cookies',
+  footerContactUs: 'Contactez-nous',
+  footerStableRelease: 'VERSION STABLE',
+  footerCraftedWithLove: 'Créé avec passion pour les joueurs du monde entier.',
+};
