+++ apps/web/src/shared/i18n/messages/home/index.ts
@@ -1,0 +1,19 @@
+import type { DeepPartial } from '../../base-types';
+import { en } from './en';
+import { es } from './es';
+import { fr } from './fr';
+import { ru } from './ru';
+import { by } from './by';
+
+export { en, es, fr, ru, by };
+
+export const homeMessages = {
+  en,
+  es,
+  fr,
+  ru,
+  by,
+} as const;
+
+/** Derived type with Partial wrapper for backward compatibility */
+export type HomeMessages = DeepPartial<typeof en>;
