+++ apps/web/src/shared/i18n/messages/home/ru.ts
@@ -1,0 +1,104 @@
+export const ru = {
+  kicker: 'Бесплатные настольные игры онлайн',
+  tagline:
+    '{{appName}} — платформа для бесплатной игры в настольные игры онлайн с друзьями: Морской бой, стратегии, карты и больше.',
+  description:
+    'Большой выбор бесплатных настольных игр онлайн. Создавайте игровые комнаты в реальном времени, приглашайте друзей, а {{appName}} возьмет на себя правила, подсчет очков и очередность ходов, чтобы вы могли сосредоточиться на веселье.',
+  primaryCtaLabel: 'Начать',
+  playWithBotsLabel: 'Играть с ботами',
+  defaultRoomName: 'Игра {{name}}',
+  heroCardBrand: 'CRITICAL',
+  heroCardPlayCta: 'Играть',
+  supportCtaLabel: 'Поддержать разработчиков',
+  downloadsTitle: 'Установить Arcadeum',
+  pwaDescription:
+    'Установите {{appName}} как веб-приложение на свое устройство для лучшего игрового опыта. Та же производительность, больше удобства.',
+  downloadsDescription: 'Загрузите последние версии прямо из веб-приложения.',
+  downloadsIosLabel: 'Скачать для iOS',
+  downloadsAndroidLabel: 'Скачать для Android',
+  // Games section
+  gamesTitle: 'Популярные игры',
+  gamesSubtitle: 'Исследуйте нашу растущую библиотеку настольных развлечений',
+  gamePlayButton: 'Играть сейчас',
+  gameTryDemo: 'Попробовать демо',
+  gameRulesButton: 'Правила игры',
+  gameComingSoon: 'Скоро',
+  gameHowToPlay: 'Как играть',
+  gameMetaPlayers: 'игроков',
+  gameMetaMatch: 'партия',
+  gameMetaPlayingNow: 'играют сейчас',
+  gamesPagerLabel: 'Карусель популярных игр',
+  gamesPagerPrev: 'Предыдущая игра',
+  gamesPagerNext: 'Следующая игра',
+  gameIncludesPacks: 'Включает 5 наборов карт:',
+  gameThemedDecks: 'Тематические колоды:',
+  gameAvailableNow: 'Доступно сейчас',
+  showMore: 'Подробнее',
+  demoBadge: 'Демо',
+  demoBadgeLabel: 'Демо-сборка',
+  showLess: 'Свернуть',
+  rulesTab: 'Правила',
+  infoTab: 'Темы игры',
+
+  // Features section
+  featuresTitle: 'Почему {{appName}}?',
+  featuresSubtitle: 'Все, что нужно для игры в настолки онлайн с друзьями',
+  featureRoomsTitle: 'Комнаты в реальном времени',
+  featureRoomsDescription:
+    'Создавайте игровые комнаты мгновенно и начинайте играть с друзьями за секунды. Загрузка не требуется.',
+  featureRulesTitle: 'Автоматизированные правила',
+  featureRulesDescription:
+    'Пусть приложение берет на себя правила, очки и управление ходами, пока вы наслаждаетесь игрой.',
+  featureCrossplatformTitle: 'Кроссплатформенность',
+  featureCrossplatformDescription:
+    'Играйте мгновенно в браузере на ПК и мобильных устройствах. Нативные приложения для iOS and Android скоро появятся.',
+  featureInviteTitle: 'Приватные комнаты и чат',
+  featureInviteDescription:
+    'Создавайте защищенные комнаты для своей группы со встроенным чатом для общения во время игры.',
+  featureSpectatorTitle: 'Режим зрителя',
+  featureSpectatorDescription:
+    'Смотрите, как друзья играют в прямом эфире на вебе. Поддержка ТВ скоро появится.',
+  featureStatsTitle: 'Статистика игры',
+  featureStatsDescription:
+    'Отслеживайте процент побед, историю и достижения во всех играх.',
+  featureTournamentsTitle: 'Турниры',
+  featureTournamentsDescription:
+    'Участвуйте в рейтинговых событиях и докажите свое мастерство против лучших игроков.',
+  // How it works section
+  howItWorksTitle: 'Как это работает',
+  howItWorksSubtitle: 'Начните за три простых шага',
+  stepCreateTitle: 'Создайте или войдите в комнату',
+  stepCreateDescription:
+    'Начните новую игровую комнату или введите код приглашения, чтобы присоединиться к существующей сессии.',
+  stepInviteTitle: 'Пригласите друзей',
+  stepInviteDescription:
+    'Поделитесь ссылкой на комнату или кодом с друзьями. Они могут присоединиться с любого устройства мгновенно.',
+  stepPlayTitle: 'Играйте вместе',
+  stepPlayDescription:
+    'Наслаждайтесь настольными играми, пока {{appName}} автоматически управляет правилами, ходами и очками.',
+  // Coming soon labels
+  comingSoon: 'Скоро',
+  // Footer
+  footerFollowUs: 'Подпишитесь на нас',
+  footerCommunity: 'Присоединяйтесь к сообществу',
+  footerRights: '© {{year}} {{appName}}. Все права защищены.',
+  footerDescription:
+    'Ваша идеальная платформа для соревновательных и казуальных настольных игр.',
+  footerPlatformTitle: 'Платформа',
+  footerAllGames: 'Все игры',
+  footerSeaBattle: 'Морской бой',
+  footerTournaments: 'Турниры',
+  footerLeaderboards: 'Таблицы лидеров',
+  footerRewards: 'Награды',
+  footerResourcesTitle: 'Ресурсы',
+  footerHelpCenter: 'Центр помощи',
+  footerGamingBlog: 'Блог о играх',
+  footerDevelopers: 'Разработчикам (API)',
+  footerLegalTitle: 'Юридическая информация',
+  footerPrivacyPolicy: 'Политика конфиденциальности',
+  footerTermsOfService: 'Условия использования',
+  footerCookiePolicy: 'Политика использования файлов cookie',
+  footerContactUs: 'Связаться с нами',
+  footerStableRelease: 'СТАБИЛЬНЫЙ ВЫПУСК',
+  footerCraftedWithLove: 'Создано с любовью для игроков со всего мира.',
+};
